refactor(web): abstract AccountDisplay
